Added version check for newer feature
Actually we have the .NOW function that is introduced
in 4.2. If you use a previous version of Mongodb you
will find wou full of Warning log from GetUtcDateTime()
method.

It would be really better to add the concept of retrieving
the version of the server when you initialize the context
and for such functions, just them only if they are supported
and use a fallback otherwise.